chore: rename comic_viewer to redViewer
