+++ docs/index.md
@@ -64,6 +64,13 @@ features:
 - 读剪贴板方式流，字如其名
 - 去重，加标识符等
 
+## 食用搭配(阅读器)
+
+完全适配 CGS 而制，取(改)了独特的名字 `RedViewer (RV)`  
+加上最近对其手撕了几十个 commit 血改，还在更新中！所以再次推上  
+
+[![点击前往redViewer](https://github-readme-stats.vercel.app/api/pin/?username=jasoneri&repo=redViewer&show_icons=true&bg_color=60,ef4057,cf4057,c44490&title_color=4df5b4&hide_border=true&icon_color=e9ede1&text_color=e9ede1)](https://github.com/jasoneri/redViewer)
+
 ## 致谢声明
 
 ### Credits
